+++ mobile/android/android-components/components/browser/state/src/main/java/mozilla/components/browser/state/state/ContentState.kt
@@ -41,6 +41,7 @@ import mozilla.components.concept.engine.window.WindowRequest
  * @property canGoForward whether or not there's an history item to navigate forward to.
  * @property webAppManifest the Web App Manifest for the currently visited page (or null).
  * @property firstContentfulPaint whether or not the first contentful paint has happened.
+ * @property pictureInPictureEnabled True if the session is being displayed in PIP mode.
  */
 data class ContentState(
     val url: String,
@@ -64,5 +65,6 @@ data class ContentState(
     val canGoForward: Boolean = false,
     val webAppManifest: WebAppManifest? = null,
     val firstContentfulPaint: Boolean = false,
-    val history: HistoryState = HistoryState()
+    val history: HistoryState = HistoryState(),
+    val pictureInPictureEnabled: Boolean = false
 )